support other primary keys

+++ entity-store.go
@@ -1,6 +1,8 @@
 package gormup
 
-import "context"
+import (
+	"context"
+)
 
 type entityStore struct {
 	store Store
@@ -13,7 +15,10 @@ func newEntityStore(store Store) *entityStore {
 }
 
 func (s *entityStore) Set(ctx context.Context, ent *entity) {
-	s.store.Set(ctx, ent.Key(), ent)
+	s.store.Set(ctx, ent.GetKey(), ent)
+	for _, key := range ent.GetOtherKeys() {
+		s.store.Set(ctx, key, ent)
+	}
 }
 
 func (s *entityStore) Get(ctx context.Context, key string) *entity {
@@ -25,6 +30,17 @@ func (s *entityStore) Get(ctx context.Context, key string) *entity {
 	return ent
 }
 
+func (s *entityStore) GetByFieldValue(ctx context.Context, table, filed, val string) *entity {
+	return s.Get(ctx, getEntityKey(table, filed, val))
+}
+
 func (s *entityStore) Delete(ctx context.Context, key string) {
+	ent := s.Get(ctx, key)
+	if ent == nil {
+		return
+	}
 	s.store.Delete(ctx, key)
+	for _, k := range ent.GetOtherKeys() {
+		s.store.Delete(ctx, k)
+	}
 }
